# Micro-optimize ActiveRecord::Core#hash